final version

+++ srcs/libft/libft.h
@@ -6,12 +6,13 @@
 /*   By: meldora <meldora@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/02 13:11:28 by meldora           #+#    #+#             */
-/*   Updated: 2021/03/15 20:57:32 by meldora          ###   ########.fr       */
+/*   Updated: 2021/03/16 12:52:39 by meldora          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
+
 # include <stdlib.h>
 # include <unistd.h>
 
@@ -71,7 +72,7 @@ t_list					*ft_lstmap(t_list *lst, void *(*f)(void *),
 						void (*del)(void *));
 
 int						ft_strcmp(const char *s1, const char *s2);
-double					ft_atof(char *s);
+float					ft_atof(char *s);
 int						ft_isspace(char c);
 
 #endif
